sp/chk precomp kernel 264 (#41)
* Input config file to find pre kernel

* code changes w.r.t prekcache test

* #267

* #static analysis report fix
        #extra cout statements removed
        #nogpuhandle else part added

        #unused variables removed
	#code CI check support.
	#main.cpp and conv_fin.hpp are modified for same.

Test precompiled kernel evaluation #264

* addressed merge issues

* fixed bad merge

+++ src/main.cpp
@@ -50,6 +50,7 @@ using json = nlohmann::json;
 
 [[gnu::noreturn]] void Usage()
 {
+
     printf("Usage: ./fin *input_json *output_json\n\n");
     printf("Supported arguments:\n");
     printf("-i *input_json\n");
@@ -149,7 +150,7 @@ int main(int argc, char* argv[], char* envp[])
             }
             else if(command["config"]["cmd"] == "bnormfp16")
             {
-                f = std::make_unique<fin::BNFin<float16, float>>(command);
+                f = std::make_unique<fin::BNFin<float, float>>(command);
             }
             else
             {
